fix: resolve CI build failures across all platforms
- Add vcpkg execution to Windows configure script with --recurse flag
- Add GLM and GLI dependencies to vcpkg-lock.json
- Enable VCPKG_MANIFEST_MODE for windows64-deploy preset
- Fix invalid cast in wwmemlog.cpp for cross-platform compatibility

Fixes #123

+++ Core/Libraries/Source/WWVegas/WWDebug/wwmemlog.cpp
@@ -392,7 +392,7 @@ ActiveCategoryStackClass::operator = (const ActiveCategoryStackClass & that)
 ***************************************************************************************************/
 ActiveCategoryStackClass & ActiveCategoryClass::Get_Active_Stack()
 {
-	int current_thread = static_cast<int>(::GetCurrentThreadId());
+	int current_thread = ::GetCurrentThreadId();
 
 	/*
 	** If we already have an allocated category stack for the current thread,
